+++ crates/orchestrator/src/data_storage/mod.rs
@@ -1,24 +1,22 @@
-mod aws_s3;
-mod types;
+pub mod aws_s3;
+pub mod types;
 
 use async_trait::async_trait;
-use aws_sdk_s3::primitives::ByteStream;
-use aws_sdk_s3::Error;
 use bytes::Bytes;
+use color_eyre::Result;
 use mockall::automock;
 
 /// DataStorage trait contains the functions used to store and get the data from
 /// the cloud provider storage.
 /// The proposed storage format is :
-/// ----s3
 ///     ----<block_number>
-///         ----<snos_output.json>
-///         ----<kzg.txt>
+///         ----<snos_output.json> (stored during the SNOS job)
+///         ----<blob_data.txt> (stored during the DA job)
 #[automock]
 #[async_trait]
 pub trait DataStorage: Send + Sync {
-    async fn get_data(&self, key: &str) -> Result<Bytes, Error>;
-    async fn put_data(&self, data: ByteStream, key: &str) -> Result<(), Error>;
+    async fn get_data(&self, key: &str) -> Result<Bytes>;
+    async fn put_data(&self, data: Bytes, key: &str) -> Result<()>;
 }
 
 /// **DataStorageConfig** : Trait method to represent the config struct needed for