+++ crates/orchestrator/src/jobs/proving_job/mod.rs
@@ -41,13 +41,14 @@ impl Job for ProvingJob {
     }
 
     async fn process_job(&self, config: &Config, job: &mut JobItem) -> Result<String> {
-        // TODO: allow to donwload PIE from S3
+        // TODO: allow to download PIE from storage
         let cairo_pie_path: PathBuf = job
             .metadata
             .get(JOB_METADATA_CAIRO_PIE_PATH_KEY)
             .map(|s| PathBuf::from_str(s))
             .ok_or_else(|| eyre!("Cairo PIE path is not specified (prover job #{})", job.internal_id))??;
-        let cairo_pie = CairoPie::read_zip_file(&cairo_pie_path).unwrap();
+        let cairo_pie = CairoPie::read_zip_file(&cairo_pie_path)
+            .expect("Not able to read the cairo PIE file from the zip file provided.");
         let external_id = config.prover_client().submit_task(Task::CairoPie(cairo_pie)).await?;
         Ok(external_id)
     }