+++ crates/orchestrator/src/tests/jobs/proving_job/mod.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use crate::config::{config, config_force_init};
 use httpmock::prelude::*;
 use prover_client_interface::{MockProverClient, TaskStatus};
 use rstest::*;
@@ -14,7 +15,7 @@ use crate::jobs::Job;
 #[rstest]
 #[tokio::test]
 async fn test_create_job() {
-    let config = init_config(None, None, None, None, None, None).await;
+    let config = init_config(None, None, None, None, None, None, None).await;
     let job = ProvingJob
         .create_job(
             &config,
@@ -40,7 +41,7 @@ async fn test_verify_job(#[from(default_job_item)] mut job_item: JobItem) {
     let mut prover_client = MockProverClient::new();
     prover_client.expect_get_task_status().times(1).returning(|_| Ok(TaskStatus::Succeeded));
 
-    let config = init_config(None, None, None, None, Some(prover_client), None).await;
+    let config = init_config(None, None, None, None, Some(prover_client), None, None).await;
     assert!(ProvingJob.verify_job(&config, &mut job_item).await.is_ok());
 }
 
@@ -52,16 +53,25 @@ async fn test_process_job() {
     let mut prover_client = MockProverClient::new();
     prover_client.expect_submit_task().times(1).returning(|_| Ok("task_id".to_string()));
 
-    let config =
-        init_config(Some(format!("http://localhost:{}", server.port())), None, None, None, Some(prover_client), None)
-            .await;
+    let config_init = init_config(
+        Some(format!("http://localhost:{}", server.port())),
+        None,
+        None,
+        None,
+        Some(prover_client),
+        None,
+        None,
+    )
+    .await;
+
+    config_force_init(config_init).await;
 
     let cairo_pie_path = format!("{}/src/tests/artifacts/fibonacci.zip", env!("CARGO_MANIFEST_DIR"));
 
     assert_eq!(
         ProvingJob
             .process_job(
-                &config,
+                config().await.as_ref(),
                 &mut JobItem {
                     id: Uuid::default(),
                     internal_id: "0".into(),