+++ crates/orchestrator/src/tests/workers/update_state/mod.rs
@@ -86,6 +86,7 @@ async fn test_update_state_worker(
         Some(da_client),
         None,
         None,
+        None,
     )
     .await;
     config_force_init(config).await;