+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Added
 
+- Function to calculate the kzg proof of x_0.
+- Tests for updating the state.
+- Function to update the state and publish blob on ethereum in state update job.
+
 ## Changed
 
 ## Removed
+
+- `fetch_from_test` argument
 
 ## Fixed